[IA-4620] create fetchLeoOnce to isolate fetches coming from the signOut flow
