+++ src/auth/auth-fetch.ts
@@ -3,19 +3,12 @@ import { FetchFn } from 'src/libs/ajax/data-client-common';
 
 import { AuthTokenState, getAuthToken, getAuthTokenFromLocalStorage, loadAuthToken, sendRetryMetric } from './auth';
 import { sessionExpirationErrorMessage } from './auth-errors';
+import { authOpts } from './auth-options';
 import { signOut, SignOutCause } from './signout/sign-out';
 
 //
 // Auth mechanics for use on typical auth-session scoped application data requests.
 //
-
-export const authOpts = (token = getAuthToken()) => ({ headers: { Authorization: `Bearer ${token}` } });
-
-export const withAuthSession =
-  (wrappedFetch: FetchFn): FetchFn =>
-  (url, options) => {
-    return wrappedFetch(url, _.merge(options, authOpts()));
-  };
 
 const isUnauthorizedResponse = (error: unknown): boolean => error instanceof Response && error.status === 401;
