+++ src/libs/ajax/DataRepo.ts
@@ -1,6 +1,6 @@
 import { jsonBody } from '@terra-ui-packages/data-client-core';
 import * as _ from 'lodash/fp';
-import { authOpts } from 'src/auth/auth-fetch';
+import { authOpts } from 'src/auth/auth-options';
 import { fetchDataRepo } from 'src/libs/ajax/ajax-common';
 
 /** API types represent the data of UI types in the format expected by the backend.
