+++ src/libs/ajax/SamResources.ts
@@ -1,6 +1,6 @@
 import { jsonBody } from '@terra-ui-packages/data-client-core';
 import _ from 'lodash/fp';
-import { authOpts } from 'src/auth/auth-fetch';
+import { authOpts } from 'src/auth/auth-options';
 import { Ajax } from 'src/libs/ajax';
 import { appIdentifier, fetchSam } from 'src/libs/ajax/ajax-common';
