+++ src/libs/ajax/fetch/fetchLeoOnce.ts
@@ -1,0 +1,5 @@
+// For Leo calls which should not trigger signout on auth failure.
+import { fetchOk, withUrlPrefix } from 'src/libs/ajax/fetch/fetch-core';
+import { getConfig } from 'src/libs/config';
+
+export const fetchLeoOnce = withUrlPrefix(`${getConfig().leoUrlRoot}/`, fetchOk);
