+++ src/libs/ajax/leonardo/Runtimes.test.ts
@@ -1,4 +1,4 @@
-import { authOpts } from 'src/auth/auth-fetch';
+import { authOpts } from 'src/auth/auth-options';
 import { fetchLeo } from 'src/libs/ajax/ajax-common';
 import { FetchFn } from 'src/libs/ajax/data-client-common';
 import { Runtimes } from 'src/libs/ajax/leonardo/Runtimes';
